move trace_level to its own module
